+++ packages/clients/dish/src/plugins/SelectionObject/SelectionObject.vue
