+++ packages/clients/dish/src/plugins/SelectionObject/locales.ts
